Fix clearing the cache for an entity from a different route

+++ server/src/api/middlewares/clearRouteCache.ts
@@ -5,17 +5,21 @@ import {
     NextFunction
 } from "express";
 
+import ParamsWithId from "#types/ParamsWithId";
 import redisClient from "#redisClient";
 import sendResponse from "#utils/http/sendResponse";
 
 const clearRouteCache: RequestHandler = async (
-    request: Request,
+    { originalUrl, params }: Request,
     response: Response,
     next: NextFunction
 ): Promise<void> => {
     try {
-        const key = redisClient.createKey(request);
-        await redisClient.delete(key);
+        if (params.id) {
+            const key = redisClient.createKey(originalUrl, params as unknown as ParamsWithId);
+            await redisClient.delete(key);
+        }
+
         sendResponse(response, response.locals.data);
     } catch (error) {
         next(error);
